workflows: Holding Pen blueprint update
* Moves some functions away from Holding Pen blueprint to
  utils (pending future separation from workflows into a
  new holding pen module).

* Optimizes some error-prone code to get the workflow
  definitions when trying to fetch formatted data from
  the object.

* Removes lots of unused/old code.

* Performs some PEP8 and PEP257 updates.
