+++ invenio/modules/workflows/templates/workflows/hp_maintable.html
@@ -68,7 +68,7 @@
             <li class="dropdown">
                 <a href="#" class="dropdown-toggle dropdown-headline" data-toggle="dropdown">Status<b class="caret"></b></a>
                 <ul class="dropdown-menu" role="menu">
-                    <li><a id="version-halted" class="version-selection" name="Nedd action">Need action</a></li>
+                    <li><a id="version-halted" class="version-selection" name="Need action">Need action</a></li>
                     <li><a id="version-completed" class="version-selection" name="Done">Done</a></li>
                     <li><a id="version-running" class="version-selection" name="In process">In process</a></li>
                     <li><a id="version-initial" class="version-selection" name="New">New</a></li>
